0: Move UFLD_V2 tests to models/demos
