+++ .github/workflows/perf-device-models-impl.yaml
@@ -162,7 +162,7 @@ jobs:
         with:
           model_name: "ufld_v2"
           commands: |
-            # pytest models/demos/ufld_v2/tests -m models_device_performance_bare_metal
+            WH_ARCH_YAML=$MAGIC_ENV pytest models/demos/ufld_v2/tests/perf -m models_device_performance_bare_metal
         if: ${{ !cancelled() && matrix.test-info.civ2-compatible && matrix.test-info.job-set1 }}
         timeout-minutes: ${{ matrix.test-info.timeout }}
