+++ models/demos/ufld_v2/runner/performant_runner_infra.py
@@ -2,43 +2,17 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-import os
 
 import torch
 from loguru import logger
 from ttnn.model_preprocessing import infer_ttnn_module_args, preprocess_model_parameters
 
 import ttnn
-from models.demos.ufld_v2.reference.ufld_v2_model import TuSimple34
+from models.demos.ufld_v2.common import load_torch_model
+from models.demos.ufld_v2.tests.pcc.test_ttnn_ufld_v2 import custom_preprocessor_whole_model
 from models.demos.ufld_v2.ttnn.ttnn_ufld_v2 import TtnnUFLDv2
 from models.utility_functions import divup, is_wormhole_b0
-from tests.ttnn.integration_tests.ufld_v2.test_ttnn_ufld_v2 import custom_preprocessor_whole_model
 from tests.ttnn.utils_for_testing import assert_with_pcc
-
-
-def load_torch_model(model_location_generator=None, use_pretrained_weight=False):
-    torch_model = TuSimple34(input_height=320, input_width=800)
-    torch_model.eval()
-    if not use_pretrained_weight:
-        return torch_model
-
-    if model_location_generator == None or "TT_GH_CI_INFRA" not in os.environ:
-        weights_path = "models/demos/ufld_v2/tusimple_res34.pth"
-        if not os.path.exists(weights_path):
-            os.system("bash models/demos/ufld_v2/weights_download.sh")
-    else:
-        weights_path = (
-            model_location_generator("vision-models/ufldv2", model_subdir="", download_if_ci_v2=True)
-            / "tusimple_res34.pth"
-        )
-
-    state_dict = torch.load(weights_path)
-    new_state_dict = {}
-    for key, value in state_dict["model"].items():
-        new_key = key.replace("model.", "res_model.")
-        new_state_dict[new_key] = value
-    torch_model.load_state_dict(new_state_dict)
-    return torch_model
 
 
 def load_ttnn_model(device, torch_model, torch_input_tensor):
@@ -81,7 +55,7 @@ class UFLDPerformanceRunnerInfra:
             if self.torch_input_tensor is None
             else self.torch_input_tensor
         )
-        self.torch_model = load_torch_model()
+        self.torch_model = load_torch_model(model_location_generator=self.model_location_generator)
         self.ttnn_ufld_v2_model = load_ttnn_model(self.device, self.torch_model, self.torch_input_tensor)
         self.torch_output_tensor_1, self.torch_output_tensor_2 = self.torch_model(self.torch_input_tensor)
