--- tests/nightly/single_card/ufld_v2/test_ufld_v2_e2e_performant.py
@@ -1,1 +1,0 @@
-../../../../models/demos/ufld_v2/tests/test_ufld_v2_e2e_performant.py
